pyproject.toml
[project]
name = "six-figure-rag-api"
version = "0.1.0"
description = ""
authors = [
    {name = "Your Name",email = "you@example.com"}
]
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "fastapi (>=0.121.2,<0.122.0)",
    "uvicorn (>=0.38.0,<0.39.0)",
    "supabase (>=2.24.0,<3.0.0)",
    "python-dotenv (>=1.2.1,<2.0.0)",
    "clerk-backend-api (>=4.0.0,<5.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
